+++ src/main/java/kitchenpos/menu/tobe/application/dto/request/CreateMenuGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.tobe.application.dto;
+package kitchenpos.menu.tobe.application.dto.request;
 
 public record CreateMenuGroupRequest(String name) {
 }
