+++ src/main/java/kitchenpos/menu/tobe/application/dto/request/CreateMenuRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.tobe.application.dto;
+package kitchenpos.menu.tobe.application.dto.request;
 
 import java.math.BigDecimal;
 import java.util.List;
